+++ tests/factories/deck.py
@@ -1,0 +1,13 @@
+import factory
+
+from spellbot.models import Deck
+
+
+class DeckFactory(factory.alchemy.SQLAlchemyModelFactory):
+    id = factory.Sequence(lambda n: 50 + n)
+    name = factory.Faker("name")
+    link = factory.Faker("url")
+
+    class Meta:
+        model = Deck
+        sqlalchemy_session_persistence = "flush"